+++ backend/shared/src/helpers/fn-queue.ts
@@ -1,6 +1,6 @@
 import { intersection, remove } from 'lodash'
 import { APIError } from 'common/api/utils'
-import { contractBetCache, revalidateBetCache } from 'shared/helpers/bet-cache'
+import { getKeys, invalidateCache } from 'shared/helpers/bet-cache'
 
 export const DEFAULT_QUEUE_TIME_LIMIT = 10_000
 
@@ -45,12 +45,12 @@ export const createFnQueue = (props?: { timeout?: number }) => {
 
   const enqueueFn = async <T>(fn: () => Promise<T>, dependencies: string[]) => {
     const res = await enqueuePrivate(fn, dependencies, false)
-    const contractIds = Object.keys(contractBetCache)
+    const contractIds = getKeys()
     // TODO I think this could work for refreshing user balances as well
     const sharedDeps = intersection(contractIds, dependencies)
     if (sharedDeps.length > 0) {
       sharedDeps.forEach((contractIdToRevalidate) => {
-        revalidateBetCache(contractIdToRevalidate)
+        invalidateCache(contractIdToRevalidate)
       })
     }
     return res
